python_scripts/requirements.txt
#/home/dean/dean/projects1/withoutbonus/python_scripts/requirements.txt
nats.py
asyncio
